Phase 2: fixed timestep, pausable GameClock, event bus, first extractions
- animate() now runs the simulation in fixed 1000/60 ms steps behind an
  accumulator, with a 250 ms frame clamp; game speed no longer tracks the
  display refresh rate.
- src/core/clock.ts: pausable GameClock (now/after/every/cancel/clear),
  ticked from the fixed step so pause freezes every gameplay timer. All raw
  gameplay setTimeouts migrated (guard reactions, alert chains, choke kill,
  backup waves, alert icon, setLastSeen); the patrol backoff counts game
  time; clearStage's clear-all-timeouts hack is one gameClock.clear().
- src/core/events.ts: the on/off/emit bus, with camera kickback as its
  first real seam ('camera:kickback' from both shoot paths).
- Strangler step: input, camera and particles systems extracted from
  main.ts into src/systems/ (main.ts 2,969 -> 2,297 lines); ~20 names
  removed from legacy-globals.d.ts and the window republish lists; dead
  scaleStageChild deleted.
- 22 new Vitest cases for clock and bus (67 total). Headless verification:
  hero speed holds 235-240 px/s from 13 to 25 FPS (pre-change build: 46.6
  px/s at 13 FPS); guard reaction, backup waves and the bomb fuse all
  freeze under pause and resume correctly.
