+++ src/legacy/main.ts
@@ -3,6 +3,11 @@ Copyright 2014,2015, Jordan O'Leary, All rights reserved.
 If you would like to copy or use my code, you may contact
 me at jdoleary@gmail.com
 /*******************************************************/
+import { gameClock } from '../core/clock';
+import { events } from '../core/events';
+import { mouseMove, addKeyHandlers, removeHandlers } from '../systems/input';
+import { updateCamera } from '../systems/camera';
+import { updateParticles, shardParticleSplatter, bloodParticleSplatter, ejectShell } from '../systems/particles';
 ////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////
 /*
@@ -24,8 +29,6 @@ window.shell2 ??= undefined;
 window.shell3 ??= undefined;
 window.shell4 ??= undefined;
 window.shell5 ??= undefined;
-window.shell_speed = 10;
-window.blood_speed = 1;
 
 window.shellTextures = [];
 window.shellType = 2;
@@ -43,14 +46,6 @@ window.debug_on = false;
 
 function getColor(x,y){
     return {r:50,g:50,b:50,a:1};
-}
-function mouseMove(e){
-    //Viewport coords, not document coords: camera.getMouse wants the position within the
-    //canvas, and the canvas sits at the top-left of an unscrolled page. pageX/pageY add
-    //the scroll offset, which silently skewed aim whenever the page could scroll at all.
-    mouse_relative.x = e.clientX;
-    mouse_relative.y = e.clientY;
-
 }
 function windowSetup(){
     //Mr Doob's Stats.js:
@@ -340,11 +335,9 @@ function clearStage(){
             button = null;
         }
     }
-    //clear all timeouts
-    var id = window.setTimeout(function() {}, 0);
-    while (id--) {
-        window.clearTimeout(id); // will do nothing if no timeout with id is present
-    }
+    //Every gameplay delay lives on the GameClock now, so a restart is one call — the
+    //old version looped over every window timeout id to kill stray setTimeouts.
+    gameClock.clear();
 
     //removeHandlers:
     console.log('clear stage');
@@ -388,7 +381,13 @@ function startGame(){
     clearStage();
     
     state = states["Gameplay"];
-    
+
+    //fresh run, fresh clock: lastTimeStamp only advances during gameplay frames, so
+    //after a stay in the menu it is stale and would count the whole menu visit as one
+    //(clamped) catch-up burst.
+    lastTimeStamp = null;
+    step_accumulator = 0;
+
     //initialize variables:
     keys = {w: false, a: false, s: false, d: false, r: false, f: false, v: false, g:false, space:false, shift:false, LMB:false, RMB:false};
     stage_child = new PIXI.Container();//replaces stage for scaling
@@ -689,6 +688,19 @@ Animate Loop
 ////////////////////////////////////////////////////////////
 window.lastTimeStamp ??= undefined;
 window.deltaTime ??= undefined;
+//Fixed timestep (roadmap Phase 2). The simulation always steps in STEP_MS increments;
+//the render loop just decides how many steps to run to catch up to wall time. Movement,
+//bullets and particles are per-tick constants, so a fixed tick rate is what makes game
+//speed independent of the display's refresh rate (144 Hz no longer runs 2.4x fast, a
+//30 FPS cap no longer runs at half speed).
+window.STEP_MS = 1000/60;
+//Longest slice of wall time one frame is allowed to simulate. Returning to a
+//backgrounded tab (rAF stops entirely) would otherwise queue thousands of catch-up
+//steps — and a machine too slow to simulate in real time would fall further behind
+//every frame trying (the classic spiral of death). Beyond this cap the game slows
+//down instead of freezing.
+window.MAX_FRAME_MS = 250;
+window.step_accumulator = 0;
 function animate(time) {
     if(state == 0){
     }else if(state == 1){
@@ -696,23 +708,27 @@ function animate(time) {
         if(!lastTimeStamp) lastTimeStamp = time;
         deltaTime = time - lastTimeStamp;
         lastTimeStamp = time;
-        //console.log('delta: ' + deltaTime);
-        //console.log('time: ' + time);
-        
-        
+        if(deltaTime > MAX_FRAME_MS)deltaTime = MAX_FRAME_MS;
+
         stats.begin();//Mr Doob's Stats.js
-        if(!pause)gameloop(deltaTime);
+        if(!pause){
+            step_accumulator += deltaTime;
+            while(step_accumulator >= STEP_MS){
+                step_accumulator -= STEP_MS;
+                gameloop(STEP_MS);
+            }
+        }
         stats.end();//Mr Doob's Stats.js
-        
-        
+
+
     }
-    
+
     // render the stage
     renderer.render(stage);
     //request another animate call
-    requestAnimationFrame(animate);	
+    requestAnimationFrame(animate);
 
-    
+
 }
 
 
@@ -835,7 +851,7 @@ function gameloop_guards(deltaTime){
                             //if guard can't shoot yet (reaction time)
                             if(!guard.reacting){
                                 guard.reacting = true;
-                                setTimeout(reactionTimeout.bind(guard), guard.shoot_speed);
+                                gameClock.after(guard.shoot_speed, reactionTimeout.bind(guard));
                             }
                         }
                         
@@ -1250,136 +1266,6 @@ function gameloop_messages_and_tooltip(deltaTime){
         }
     }
 }
-function gameloop_zoom_and_camera(deltaTime){
-    //////////////////////
-    //Camera
-    //////////////////////
-    
- 
-    //////////////////////
-    //Zoom / Scale
-    //////////////////////
-    //this code allows the zoom / scale to change smoothly based on the mouse wheel input
-    
-   if(stage_child.scale.x < zoom - 0.05){//the 0.05 is close enough to desired value to stop so the zoom doesn't bounce back and forth.
-        stage_child.scale.x += zoom_magnitude;
-        stage_child.scale.y += zoom_magnitude;
-        changeFontSizes();
-    }else if(stage_child.scale.x > zoom + 0.05){//the 0.05 is close enough to desired value to stop so the zoom doesn't bounce back and forth.
-        stage_child.scale.x -= zoom_magnitude;
-        stage_child.scale.y -= zoom_magnitude;
-        changeFontSizes();
-    
-    }
-    
-    //loose camera
-    camera.x = hero.x + (mouse.x - hero.x)/look_sensitivity;
-    camera.y = hero.y + (mouse.y - hero.y)/look_sensitivity;
-    //don't let camera show out of bounds:
-    var cam_width = window_properties.width*(1/stage_child.scale.x);
-    var cam_height = window_properties.height*(1/stage_child.scale.y);
-    var cam_adjust_x = camera.x;
-    var cam_adjust_y = camera.y;
-    
-    
-    if(camera.x < 0+cam_width/2){
-        cam_adjust_x = 0+cam_width/2;
-    }
-    if(camera.y < 0+cam_height/2){
-        cam_adjust_y = 0+cam_height/2;
-    }
-    
-    if(camera.x >= grid_width-cam_width/2){
-        cam_adjust_x = grid_width-cam_width/2;
-    }
-    if(camera.y >= grid_height-cam_height/2){
-        cam_adjust_y = grid_height-cam_height/2;
-    }
-    
-    //check both:
-    if(cam_width > grid_width){
-        //if both out of left and right limit, put camera in middle
-        cam_adjust_x = grid_width/2;
-    }
-    if(cam_height > grid_height){
-        //if both out of top and bottom limit, put camera in middle
-        cam_adjust_y = grid_height/2;
-    }
-    
-    camera.x = cam_adjust_x;
-    camera.y = cam_adjust_y;
-    
-    
-    if(camera.shaking){
-        camera.posBeforeShakex = cam_adjust_x;
-        camera.posBeforeShakey = cam_adjust_y;    
-    }
-    camera.shake();
-    
-    
-    
-    /*
-    //The below commented block is for smooth camera
-    //press space to look around
-    if(keys['space']){
-        //camera floats between hero and mouse
-        //smooth camera
-        camera.following = true;
-        camera.target = {x: hero.x + (mouse.x - hero.x)/2, y: hero.y + (mouse.y - hero.y)/2};
-    }else{
-        //set camera target to hero
-        if(camera.following){
-            //return to hero
-            camera.target = hero;
-        }else{
-            //stick to hero
-            camera.x = hero.x; 
-            camera.y = hero.y;
-        }
-    }
-    //if the camera is set to following, move it to target, otherwise, don't because it is sticking to target.
-    if(camera.following && camera.move_to_target()){
-        if(camera.target == hero){
-            camera.following = false;//when camera reaches it's target, turn off following so it can just stick.
-        }
-    }*/
-    //set the stage_child to the correct position, the stage_child now acts as the camera.
-    //stage_child.x = (-camera.x+cam_width/2)*stage_child.scale.x;
-    //stage_child.y = (-camera.y+cam_height/2)*stage_child.scale.y;
-    //camera with kickback:
-    if(stage_child.kickx != null && stage_child.kicky != null){
-        var movement = moveToTarget(stage_child.kickx,stage_child.kicky,camera.x,camera.y,kickback_speed);
-        //set kickback to null if it reaches its target:
-        if(movement.x == camera.x && movement.y == camera.y){
-            stage_child.kickx = null;
-            stage_child.kicky = null;
-        }else{
-            //move kickback to where it should be:
-            stage_child.kickx = movement.x;
-            stage_child.kicky = movement.y;
-            stage_child.x = (-stage_child.kickx+cam_width/2)*stage_child.scale.x;
-            stage_child.y = (-stage_child.kicky+cam_height/2)*stage_child.scale.y;
-        }
-    }else{
-        //camera without kickback
-        stage_child.x = (-camera.x+cam_width/2)*stage_child.scale.x;
-        stage_child.y = (-camera.y+cam_height/2)*stage_child.scale.y;
-        
-    }
-    
-    
-}
-function scaleStageChild(a){
-    stage_child.scale.x = a;
-    stage_child.scale.y = a;
-    console.log(stage_child.x + "," + stage_child.y);
-    var cam_width = window_properties.width*(1/stage_child.scale.x);
-    var cam_height = window_properties.height*(1/stage_child.scale.y);
-    console.log(cam_width + "," + cam_height);
-}
-function changeFontSizes(){
-    tooltip.style.font = 30/stage_child.scale.x + "px Arial";      
-}
 function gameloop_getawaycar_and_loot(deltaTime){
     //////////////////////
     //Getaway Car and Loot
@@ -1737,6 +1623,12 @@ function make_starburst_without_limit(unit){
 
 function gameloop(deltaTime){
     //////////////////////
+    //advance gameplay timers (guard reactions, chokes, backup waves, ...).
+    //Ticked from the fixed step, so pause freezes every pending timer for free.
+    //////////////////////
+    gameClock.update(deltaTime);
+
+    //////////////////////
     //update Mouse
     //////////////////////
     if(mouse_relative.x != -10000)mouse = camera.getMouse(mouse_relative);//only set mouse position if the mouse is on the stage
@@ -1769,7 +1661,7 @@ function gameloop(deltaTime){
             hero.gun.ammo--;
             doGunShotEffects(hero, hero.gun.silenced);//plays sound and shows affects
             //kickback camera
-            kickback();
+            events.emit('camera:kickback');
             ejectShell(hero);
             hero.shoot();
             if(!hero.gun.silenced)unsilenced_gun();//make noise (not real sound, but noise for guards) which draws guards
@@ -1851,39 +1743,8 @@ function gameloop(deltaTime){
     //////////////////////
     //update particles
     //////////////////////
-    for(var i = 0; i < shells.length; i++){
-        if(tickParticle(shells[i],20,true)){
-            shells.splice(i,1);
-            i--;
-        }
-    }
-    for(var i = 0; i < shards.length; i++){
-        if(tickParticle(shards[i],20,true)){
-            shards.splice(i,1);
-            i--;
-        }
-    }
-    for(var i = 0; i < bloods.length; i++){
-        var blood = bloods[i];
-        //shrink blood particle:
-        blood.scale.x*=0.7;
-        blood.scale.y*=0.7;
-        
-        
-        var blood_x_mod = randomFloatWithBias2(-10,10);
-        var blood_y_mod = randomFloatWithBias2(-10,10);
-        var blood_size_mod = randomFloatWithBias2(1,blood.scale.y*20);
-        //var skip_blood_draw = randomIntFromInterval(0,3);
-        drawBloodTrail(blood.position.x+blood_x_mod,blood.position.y+blood_y_mod,blood_size_mod);
-        
-        //remove when done ticking
-        if(tickParticle(blood,7,false)){
-            bloods.splice(i,1);
-            i--;
-        }
-        
-    }
-    
+    updateParticles(deltaTime);
+
 
     
     
@@ -1983,7 +1844,7 @@ function gameloop(deltaTime){
         
     }
     
-    gameloop_zoom_and_camera(deltaTime);
+    updateCamera(deltaTime);
     //causing slowdown?
     if(debug_on)updateDebugInfo();
     
@@ -2034,408 +1895,6 @@ function updateDebugInfo(){
         "corner: " + screenCorner.x + "," + screenCorner.y + "<br>"
     );
 }
-//Controls the behavior of a particle during one game tick
-function tickParticle(particle,tick_max,bounce){
-    //particle should have .dy .dx .dr. and tick
-    particle.position.y += particle.dy;
-    //check the y to see if it has gone into a wall
-    if(grid.isWallSolid_coords(particle.position.x,particle.position.y)){
-        particle.position.y -= particle.dy*2;
-        particle.dy *= -1;
-    }
-    particle.position.x -= particle.dx;
-        //check the x to see if it has gone into a wall
-        if(grid.isWallSolid_coords(particle.position.x,particle.position.y)){
-            // if particle should bounce off walls
-            if(bounce){
-                    particle.position.x += particle.dx*2;
-                    particle.dx *= -1;
-            }else{
-                return true;
-            }
-        }
-    
-    particle.dx *= 0.9;
-    particle.dy *= 0.9;
-    particle.rotation += particle.dr;
-    particle.tick++;
-    //remove from array once it is done moving (OPTIMIZATION)
-    if(particle.tick > tick_max){
-        return true;//remove it from array
-    }
-    return false;
-}
-////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////
-/*
-Key Handlers
-*/
-////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////
-
-function addKeyHandlers(){
-    //override right click:
-    if (document.addEventListener) {
-        document.addEventListener('contextmenu', function(e){
-            e.preventDefault();
-        }, false);
-    } else {
-        (document as any).attachEvent('oncontextmenu', function() {
-            window.event.returnValue = false;
-        });
-    }
-    window.onkeydown = function(e){
-        //this function is called every frame that said key is down
-        var code = e.keyCode ? e.keyCode : e.which;
-        //keyinfo[code] = String.fromCharCode(code);
-        if(hero.alive){
-            /*
-            if(code == 49){hero.changeGun(0);}//key 1
-            if(code == 50){hero.changeGun(1);}//key 2
-            if(code == 51){hero.changeGun(2);}//key 3
-            if(code == 52){hero.changeGun(3);}//key 4
-            if(code == 53){hero.changeGun(4);}//key 5
-            if(code == 54){hero.changeGun(5);}//key 6
-            */
-            if(code == 87){keys['w'] = true;}
-            if(code == 65){keys['a'] = true;}
-            if(code == 83){keys['s'] = true;}
-            if(code == 68){keys['d'] = true;}
-            if(code == 80){
-                //key p
-                hero.spyglass_equipped = !hero.spyglass_equipped;
-                
-                if(hero.spyglass_equipped){
-                    hero.gunOut = false;
-                    if(!hero.gunOut)hero.gun_shot_line.graphics.clear();
-                    setHeroImage();
-                }
-            
-            }
-            if(code == 71){
-                // !keys['g'] makes it so that it will only be called once for a single press of the letter
-                if(!keys['g']){
-                    hero.gunOut = !hero.gunOut;
-                    if(!hero.gunOut)hero.gun_shot_line.graphics.clear();
-                    setHeroImage();
-                    
-                    if(hero.gunOut){
-                        hero.spyglass_equipped = false;
-                    }
-                }
-                keys['g'] = true;
-            }
-            /*if(code == 80){
-                //'p'
-                pause = !pause;
-            }*/
-            if(code == 82){
-                keys['r'] = true;
-                hero.reload();
-            }
-            if(code == 70){
-                //plant bomb
-                if(!keys['f'] && !bomb.sprite.visible){
-                
-                    if(hero.ability_remote_bomb){
-                        //remote bomb
-                        hero.moving = false;
-                        circProgBar.reset(hero.x,hero.y,1500,function(){
-                            plantBomb();
-                            bomb_tooltip.text = ("Press 'f' to detonate");
-                        });
-                        bombs_left--;
-                    }else if(hero.ability_timed_bomb){
-                        //timed bomb
-                        //if f isn't already pressed and bomb isn't already set
-                        if(bombs_left>0){
-                            hero.moving = false;
-                            circProgBar.reset(hero.x,hero.y,1500,function(){plantBomb();setBomb(5000);});
-                            bombs_left--;
-                        }else{
-                            newFloatingMessage("No Bombs Left",{x:hero.x,y:hero.y},"#FFaa00");
-                        }
-                    }
-                     
-                }
-                if(!keys['f'] && bomb.sprite.visible){
-                    if(hero.ability_remote_bomb){
-                        //set remote bomb
-                        setBomb(500);
-                    }
-                }
-                keys['f'] = true;
-            
-            }
-            if(code == 86){
-                // !keys['v'] makes it so that it will only be called once for a single press of the letter
-                if(!keys['v']){
-                    circProgBar.heroMaskProg(hero.ability_toggle_mask_speed,useMask,!hero.masked);
-                    
-                    
-                }
-                keys['v'] = true;
-            }
-            if(code == 16){
-                keys['shift'] = true;
-                //cannot sprint while dragging something
-                if(!hero_drag_target){
-                    hero.speed = hero.speed_sprint;
-                }
-            
-            }
-            if(code == 32){
-                keys['space'] = true;
-                if(!hero_drag_target){
-                    if(!grid.a_door_is_being_unlocked){
-                            //lockpick door:
-                    
-                        for(var i = 0; i < grid.doors.length; i++){
-                            var door = grid.doors[i];
-                            if(door.solid && get_distance(hero.x,hero.y,door.x+grid.cell_size/2,door.y+grid.cell_size/2) <= hero.radius*5){
-                                //if door isn't solid, then it is already unlocked.
-                                grid.a_door_is_being_unlocked = true;
-                                
-                                //timer
-                                var unlockTimeRemaining = hero.lockpick_speed;
-                                hero.lockpicking = true;
-                                circProgBar.reset(door.x+grid.cell_size/2,door.y+grid.cell_size/2,unlockTimeRemaining,grid.door_sprites[i].unlock.bind(grid.door_sprites[i]));
-                                //cancel unlocking if hero moves away from door, unless hero has unlocked remote unlock
-                                if(!hero.ability_remote_lockpick)circProgBar.distanceCancelTarget = {x:door.x,y:door.y};
-                                
-                                return;//unlocking doors succeeds loot interactions.  (Hero can unlock door while holding loot).
-                            }
-                        }
-                    
-                        //check if any dead guards are close enough to be dragged.
-                        for(var i = 0; i < guards.length; i++){
-                            var guard = guards[i];
-                            if(get_distance(hero.x,hero.y,guard.x,guard.y) <= hero.radius*dragDistance){
-                            
-                                //check if any dead guards are close enough to be dragged.
-                                if(!guard.alive){
-                                    //hero is dragging a dead body
-                                    
-                                    //slow down hero speed because he just started dragging something.
-                                    hero.speed = hero.speed/2;
-                                    hero_drag_target = guard;
-                                    hero_drag_target.speed = hero.speed;
-                                    hero_drag_target.stop_distance = hero.radius*2;//I don't know why but the stop distance here seems to need to be bigger by a factor of 10
-                                    //return;//don't return, this allows choking out a guard to have higher precedence than dragging a body
-                                }else{
-                                    //hero is choking out a live guard who is not already alarmed:
-                                    newMessage('You are choking out a guard!');
-                                    //play_sound(sound_guard_choke);
-                                            
-                                            
-                                    //add to stats:
-                                    jo_store_inc("guardsChoked");
-                                    
-                                    guard.moving = true;
-                                    guard.path = [];
-                                    guard.target = {x: null, y:null}; 
-                                    guard.being_choked_out = true;
-                                    //slow down hero speed because he just started dragging something.
-                                    hero.speed = hero.speed_walk/2;
-                                    hero_drag_target = guard;
-                                    hero_drag_target.speed = hero.speed;
-                                    hero_drag_target.stop_distance = hero.radius*2;//I don't know why but the stop distance here seems to need to be bigger by a factor of 10
-                                    setTimeout(function(){
-                                        //check that the guard is still being choked out, if not, he's not dead so don't kill() him
-                                        if(hero_drag_target == this){
-                                            newMessage('The guard is dispached!');
-                                            this.kill();
-                                            //if space isn't still being held release body:
-                                            if(!keys['space']){
-                                                //drag is a toggle action so release current drag target.
-                                                hero_drag_target.stop_dragging();
-                                                hero_drag_target = null;
-                                                //bring hero speed back to normal
-                                                hero.speed = hero.speed_walk;
-                                            }
-                                        }
-                                    }.bind(guard), hero.ability_choke_speed);
-                                    return;
-                                }
-
-                            }
-                            
-                                
-                            
-                        }
-                        
-                        //check if hero is close enough to bug a camera:
-                        for(var s = 0; s < security_cameras.length; s++){           
-                            var cam = security_cameras[s];
-                            if(hero.alive && get_distance(hero.x,hero.y,cam.x,cam.y) <= hero.radius*dragDistance){
-                                cam.hacked = true;
-                            }
-                        }
-                    
-                        //note: dragging guards takes precedence over all the following actions.
-                        
-                        //check if hero is close enough to the security camera computer to disable cameras:
-                        if(!cameras_disabled && get_distance(hero.x,hero.y,computer_for_security_cameras.x,computer_for_security_cameras .y) <= hero.radius*4){
-                            cameras_disabled = true;
-                            newMessage('All security cameras have been disabled!');
-                            computer_for_security_cameras.sprite.texture = (img_computer_off);
-                            for(var i = 0; i < security_cameras.length; i++){
-                                security_cameras[i].sprite.texture = (img_cam_off);
-                            
-                            }
-                        }
-                    }
-                    
-                }
-                
-            }
-        }
-        
-        if(code == 27){
-            //esc
-            startMenu();
-        }
-        
-        hero_move_animation_check();
-    };
-    window.onkeyup = function(e){
-        var code = e.keyCode ? e.keyCode : e.which;
-        if(code == 87){keys['w'] = false;}
-        if(code == 65){keys['a'] = false;}
-        if(code == 83){keys['s'] = false;}
-        if(code == 68){keys['d'] = false;}
-        if(code == 70){keys['f'] = false;}
-        if(code == 71){keys['g'] = false;}
-        if(code == 82){keys['r'] = false;}
-        if(code == 86){
-            //on release of key only
-            //if(keys['v'])circProgBar.stop();//stop putting on mask
-            keys['v'] = false;
-        }
-        if(code == 16){
-            keys['shift'] = false;
-            if(hero_drag_target==null){
-                hero.speed = hero.speed_walk;
-            }
-        }
-        if(code == 32){
-            keys['space'] = false;
-            //if hero was dragging something, drop it. (Don't drop a guard while he's being choked
-            if(hero_drag_target && !hero_drag_target.alive){
-                //drag is a toggle action so release current drag target.
-                hero_drag_target.stop_dragging();
-                hero_drag_target = null;
-                //bring hero speed back to normal
-                hero.speed = hero.speed_walk;
-            }
-            //allow user to abort unlocking door:
-            if(grid.a_door_is_being_unlocked && !hero.ability_remote_lockpick){
-                circProgBar.stop();
-                hero.lockpicking = false;
-            }
-            
-            grid.a_door_is_being_unlocked = false;//unlocking stops when space is released
-        }
-        hero_move_animation_check();
-        
-    };
-    // IE9, Chrome, Safari, Opera
-    window.addEventListener("mousewheel", mouseWheelHandler, false);
-    // Firefox
-    window.addEventListener("DOMMouseScroll", mouseWheelHandler, false);
-
-    onmousedown = function(e){
-        clickEvent = e;
-        if(clickEvent.which === 1){
-            //LMB
-            if(hero_drag_target){
-                newFloatingMessage("You cannot shoot while dragging a body!",{x:hero.x,y:hero.y},"#FFaa00");
-                return;
-            }
-            keys['LMB'] = true;
-            //hero can only shoot if gun is out
-            if(hero.gunOut){
-                    //very minor camera shake:
-                if(hero.gun.ammo > 0)camera.startShake(10,0);
-                
-                if(!hero.gun.automatic){
-                    if(hero.gun.ammo > 0){
-                        //very minor camera shake:
-                        camera.shakeDecay = 1.5;
-                    
-                        hero.gun.ammo--;
-                        //newFloatingMessage("Ammo: " + hero.gun.ammo + "/6",{x:hero.x,y:hero.y},"#FFaa00");
-                        doGunShotEffects(hero, hero.gun.silenced);//plays sound and shows affects
-                        //kickback camera
-                        kickback();
-                        ejectShell(hero);
-                        
-                        hero.shoot();
-                        if(!hero.gun.silenced)unsilenced_gun();//make noise (not real sound, but noise for guards) which draws guards
-                        window.mouse_click_obj = camera.objectivePoint_ignore_shake(clickEvent);  //uses clickEvent's .x and .y to find objective click
-                        
-                        
-                    }
-                }
-                if(hero.gun.ammo<=0){
-                    newFloatingMessage("Press 'r' to reload!",{x:hero.x,y:hero.y},"#FF0000");
-                    play_sound(sound_dry_fire);
-                }
-            }
-        }else if(clickEvent.which === 3){
-            //RMB
-            keys['RMB'] = true;
-            var oldgun = hero.gun;
-            for(var i = 0; i < gun_drops.length; i++){
-                var gun_drop = gun_drops[i];
-                //if close to gun_drop, pick it up:
-                if(get_distance(hero.x,hero.y,gun_drop.x,gun_drop.y) <= hero.radius*dragDistance){
-                    pickUpGunDrop(gun_drop);
-                    break;
-                }
-            }
-            if(oldgun != hero.gun){
-              // if hero picked up a gun:
-              drop_gun(oldgun,hero.x,hero.y);
-              setHeroImage(); 
-            }
-        }
-
-    }
-    onmouseup = function(e){
-        clickEvent = e;
-        if(clickEvent.which === 1){
-            keys['LMB'] = false;
-            //set shakeDecay so that when automatic gun is done firing it will stop the shake.
-            camera.shakeDecay = 1.5;
-        }else if(clickEvent.which === 3){
-            //RMB
-            keys['RMB'] = false;
-        }
-    }
-}
-function mouseWheelHandler(e){
-    // cross-browser wheel delta
-    var e = window.event || e; // old IE support
-    var delta = Math.max(-1, Math.min(1, (e.wheelDelta || -e.detail)));
-    
-    //limit amount that cam can zoom out
-    if(delta < 0 && zoom > 0.1){
-        zoom += delta * 0.05;
-    }else if (delta >0){
-        zoom += delta * 0.05;
-    }
-}
-function removeHandlers(excludeKeyHandlers?){
-    keys = {w: false, a: false, s: false, d: false, r: false, f: false, v: false, g:false, space:false, shift:false, LMB:false};
-    
-    //if excludeKeyDown is true, don't remove the onkeydown and onkeyup listeners
-    if(!excludeKeyHandlers)window.onkeydown = null;
-    if(!excludeKeyHandlers)window.onkeyup = null;
-    window.removeEventListener("mousewheel",mouseWheelHandler);
-    window.removeEventListener("DOMMouseScroll",mouseWheelHandler);
-    onmousedown = null;
-}
 ////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////
 /*
@@ -2469,7 +1928,7 @@ function spawn_backup(){
     newMessage("The police have arrived!");
     
     for(var backup = 0; backup < numOfBackupGuards; backup++){
-        setTimeout(spawn_individual_backup,1000*backup);//wait an extra second for each guard
+        gameClock.after(1000*backup, spawn_individual_backup);//wait an extra second for each guard
     }
 }
 function spawn_individual_backup(){
@@ -2602,20 +2061,6 @@ function doGunShotEffects(unit, silenced){
         play_sound_many(sound_gun_shots);
     }
 }
-window.hero_moving = false;
-function hero_move_animation_check(){
-        var hero_was = hero_moving;
-        if(keys['w'] || keys['a'] || keys['s'] || keys['d'])hero_moving = true;
-        else hero_moving = false;
-        if(hero_moving && !hero_was){
-            hero.feet_clip.gotoAndPlay(0);
-            hero.sprite_animate = true;
-        }
-        if(!hero_moving){
-            hero.feet_clip.gotoAndStop(0);
-            hero.sprite_animate = false;
-        }
-}
 
 //show alert icon
 function set_latestAlert(unit){
@@ -2628,11 +2073,11 @@ function set_latestAlert(unit){
             alert_clip.sprite.gotoAndPlay(0);
             
             //turn off the alert after 3.5 second
-            setTimeout(function(){
-                
+            gameClock.after(3500, function(){
+
                 alert_clip.sprite.visible = false;
 
-            }, 3500);
+            });
         }
     }
     
@@ -2794,118 +2239,6 @@ window.onresize = function (event){
 
 
 }
-window.shard_limit = 2000;
-window.shardType = 0;
-//Shart images inited at images_from_sheet.js
-function shardParticleSplatter(angle,target){
-    var shardAmount = randomIntFromInterval(6,30);
-        angle += Math.PI/2;//I don't know why it's off by Pi/2 but it is.
-    for(var i = 0; i < shardAmount; i++){
-        //make new bunnies
-        window.shard = new PIXI.Sprite(currentShard);
-        //shard = new PIXI.Sprite(img_shell);
-        
-        
-        shard.anchor.x = 0.5;
-        shard.anchor.y = 0.5;
-        shard.position.x = target.x;
-        shard.position.y = target.y;
-        var randScale = randomFloatFromInterval(0.3,1);
-        shard.scale.x = randScale;
-        shard.scale.y = randScale;
-        var randSpeed = randomFloatWithBias(0.1,shell_speed*2);
-        var randRotationOffset = randomFloatFromInterval(-Math.PI/6,Math.PI/6);
-        shard.dr = randomFloatFromInterval(-0.3,0.3);//change in rotation
-        shard.dx = randSpeed*Math.sin(angle+randRotationOffset);
-        shard.dy = randSpeed*Math.cos(angle+randRotationOffset);
-        shard.tick = 0;//the amount of times that it has moved;
-        shard.rotation = (angle);
-
-        shards.push(shard);
-        particle_container.addChild(shard);
-        
-        //rotate to next image:
-        shardType++
-        shardType %= shardImages.length;
-        window.currentShard = shardImages[shardType];
-        
-        if(shards.length > shard_limit)return;
-    }
-    
-}
-function bloodParticleSplatter(angle,target){
-    //var bloodAmount = randomIntFromInterval(15,30);
-    var bloodAmount = randomIntFromInterval(30,60);
-    angle += Math.PI/2;//I don't know why it's off by Pi/2 but it is.    
-    var bloodSplat;
-    for(var i = 0; i < bloodAmount; i++){
-        //make new bunnies
-        bloodSplat = new PIXI.Sprite();
-        
-        
-        bloodSplat.anchor.x = 0.5;
-        bloodSplat.anchor.y = 0.5;
-        /*var randScale = randomFloatFromInterval(1,2);
-        bloodSplat.scale.x = randScale;
-        bloodSplat.scale.y = randScale;*/
-        var randSpeed = randomFloatWithBias(0.7,blood_speed);
-        var randRotationOffset = randomFloatWithBias(0,Math.PI/4);
-        var negativeRotationOffset = randomIntFromInterval(0,2);
-        if(negativeRotationOffset)randRotationOffset *= -1;
-        bloodSplat.dr = randomFloatFromInterval(-0.3,0.3);//change in rotation
-        bloodSplat.dx = -randSpeed*Math.sin(angle+randRotationOffset)*15;
-        bloodSplat.dy = -randSpeed*Math.cos(angle+randRotationOffset)*15;
-        //start the blood off a little away from target
-        bloodSplat.position.x = target.x+Math.sin(angle)*20;
-        bloodSplat.position.y = target.y+Math.cos(angle)*20;
-        bloodSplat.tick = 0;//the amount of times that it has moved;
-        bloodSplat.rotation = (angle);
-
-        bloods.push(bloodSplat);
-    }
-    
-}
-function ejectShell(source){
-    
-    //make new bunnies
-    window.shell = new PIXI.Sprite(img_shell);
-    
-    
-    shell.anchor.x = 0.5;
-    shell.anchor.y = 0.5;
-    shell.position.x = source.x;
-    shell.position.y = source.y;
-    shell.scale.x = 0.5;
-    shell.scale.y = 0.5;
-    var randSpeed = randomFloatWithBias(shell_speed*0.6,shell_speed*2);
-    var randRotationOffset = randomFloatFromInterval(-Math.PI/6,Math.PI/6);
-    shell.dr = randomFloatFromInterval(-0.3,0.3);//change in rotation
-    shell.dx = randSpeed*Math.sin(source.sprite.rotation+randRotationOffset);
-    shell.dy = randSpeed*Math.cos(source.sprite.rotation+randRotationOffset);
-    shell.tick = 0;//the amount of times that it has moved;
-    shell.rotation = (source.sprite.rotation);
-
-    shells.push(shell);
-    particle_container.addChild(shell);
-    //cycle shell texture
-	/*shellType++
-	shellType %= 5;
-	currentTexture = shellTextures[shellType];*/
-    //
-}
-window.kickback_speed = 5;
-window.kickback_amount = 30;
-function kickback(){
-    var d = get_distance(hero.x,hero.y,mouse.x,mouse.y);
-    var kickback_mod = randomFloatFromInterval(0,20);
-    var c = kickback_amount+kickback_mod;
-    var xx = -(c/d)*(mouse.x-hero.x);
-    var yy = -(c/d)*(mouse.y - hero.y);
-    //set stage_child kickx so that the camera will kick back
-    stage_child.kickx = xx+camera.x;
-    stage_child.kicky = yy+camera.y;
-
-}
 /*Get map from server:*/
 window.map_json = "";
 function getMapInfo(subdir, fileName){
@@ -2959,6 +2292,6 @@ function getMapInfo(subdir, fileName){
 // `window`. It is an ES module now, so the functions below are republished as
 // globals for the not-yet-extracted code that still reads them by bare name.
 // See src/legacy-bridge.ts. Each extraction deletes another line from here.
-Object.assign(window, { getColor, mouseMove, windowSetup, fullscreen, drawBloodTrail, bakeBloodTrail, getUrlVars, removeAllChildren, clearStage, startMenu, startGame, setup_map, animate, reactionTimeout, gameloop_guards, gameloop_security_cams, gameloop_bullets, gameloop_doors, gameloop_dragtarget, gameloop_messages_and_tooltip, gameloop_zoom_and_camera, scaleStageChild, changeFontSizes, gameloop_getawaycar_and_loot, gameloop_alert_animation, pickUpGunDrop, make_starburst, make_starburst_with_modified_view, make_starburst_without_limit, gameloop, updateDebugInfo, tickParticle, addKeyHandlers, mouseWheelHandler, removeHandlers, newMessage, newFloatingMessage, updateMessage, spawn_backup, spawn_individual_backup, alert_all_guards, unsilenced_gun, setHeroImage, useMask, hero_is_dead, doGunShotEffects, hero_move_animation_check, set_latestAlert, setBomb, gameloop_bomb, explodeBomb, plantBomb, drop_gun, killHero, shardParticleSplatter, bloodParticleSplatter, ejectShell, kickback, getMapInfo });
+Object.assign(window, { getColor, windowSetup, fullscreen, drawBloodTrail, bakeBloodTrail, getUrlVars, removeAllChildren, clearStage, startMenu, startGame, setup_map, animate, reactionTimeout, gameloop_guards, gameloop_security_cams, gameloop_bullets, gameloop_doors, gameloop_dragtarget, gameloop_messages_and_tooltip, gameloop_getawaycar_and_loot, gameloop_alert_animation, pickUpGunDrop, make_starburst, make_starburst_with_modified_view, make_starburst_without_limit, gameloop, updateDebugInfo, newMessage, newFloatingMessage, updateMessage, spawn_backup, spawn_individual_backup, alert_all_guards, unsilenced_gun, setHeroImage, useMask, hero_is_dead, doGunShotEffects, set_latestAlert, setBomb, gameloop_bomb, explodeBomb, plantBomb, drop_gun, killHero, getMapInfo });
 
 export {};
