+++ src/legacy/sprite_hero.ts
@@ -3,6 +3,9 @@ Copyright 2014,2015, Jordan O'Leary, All rights reserved.
 If you would like to copy or use my code, you may contact
 me at jdoleary@gmail.com
 /*******************************************************/
+import { gameClock } from '../core/clock';
+import { removeHandlers } from '../systems/input';
+import { bloodParticleSplatter } from '../systems/particles';
 function sprite_hero_wrapper(pixiSprite,speed_walk,speed_sprint){
     function sprite_hero(){
         this.speed_walk = speed_walk;
@@ -128,7 +131,7 @@ function sprite_hero_wrapper(pixiSprite,speed_walk,speed_sprint){
         this.setLastSeen = function(observer){
             if(observer){
                 //if the observer is still alive after 2 seconds and not being choked out, alert the others
-                setTimeout(function(){
+                gameClock.after(2000, function(){
                     if(observer.alive && !observer.being_choked_out){
                         if(this.lastSeenX != observer.sawHeroLastAt.x && this.lastSeenY != observer.sawHeroLastAt.y){
                             this.lastSeenX = observer.sawHeroLastAt.x;
@@ -140,7 +143,7 @@ function sprite_hero_wrapper(pixiSprite,speed_walk,speed_sprint){
                             //newMessage("Last seen " + observer.sawHeroLastAt.x + "," + observer.sawHeroLastAt.y);
                         }
                     };
-                }.bind(this), 2000);
+                }.bind(this));
             }else{
                 //if observer is null, everyone is notified immediately (gunshot or camera or something).
                 if(this.lastSeenX != this.x && this.lastSeenY != this.y){
